\docType{data}
\name{AF_PER_CFS_DAY}
\alias{AF_PER_CFS_DAY}
\title{hydrologic constants}
\format{\preformatted{ num 1.98
}}
\usage{
AF_PER_CFS_DAY
}
\description{
used for converting between flows in (k)cfs and volumes in
(K)AF
}
\note{
an exact conversion as cubic feet per acre-foot is 43560,
seconds in a day is 86400, simplifing the conversion gets
24/12.1
}
\keyword{datasets}

