% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssrip2.r
\name{dssConfig}
\alias{dssConfig}
\title{dssrip read dss configuration file}
\usage{
dssConfig(
  configFileName = options()[["dss_config_filename"]],
  defaultConfig = options()[["dss_default_config"]],
  allowedStates = options()[["dss_allowed_states"]],
  dssOverrideLocation = options()[["dss_override_location"]]
)
}
\arguments{
\item{configFileName}{defaults to package jar_config.json, otherwise can point to set of libraries from `options()`'s `dss_config_filename`}

\item{defaultConfig}{name of configuration to use if you know _exactly_ which one you want, defaults to option `dss_default_config`}

\item{allowedStates}{only allow loading of configs in these states, defaults to `tested` or `dss_allowed_states` option}

\item{dssOverrideLocation}{only allow dss libraries from this directory, will try all the configs at this location. can be set by option `dss_override_location`}
}
\description{
Reads dss configuration file from .json internal to package or externally selected by user.  Useful for managing multiple dss versions for linking to dssrip.
}
\details{
Reads the jar_config.json file to find DSS libraries for use by dssrip.
}
\author{
Evan Heisman
}
