\name{fullPathByRegex}
\alias{fullPathByRegex}
\title{fullPathByRegex Search full paths with regex.}
\usage{
fullPathByRegex(paths, pattern)
}
\value{
stuff
}
\description{
Searches full paths using regular expressions, e.g.
"/A/B/C/.*/.*/F/"
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

