\name{fullPathByWildcard}
\alias{fullPathByWildcard}
\title{fullPathByWildcard Search paths by wildcard.}
\usage{
fullPathByWildcard(paths, pattern)
}
\value{
stuff
}
\description{
Searches full paths by wildcard, e.g. "/A/B/C/*/*/F/"
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

