\name{getAllPaths}
\alias{getAllPaths}
\title{getAllPaths Returns all paths in DSS file.}
\usage{
getAllPaths(file, rebuild = FALSE)
}
\arguments{
  \item{file}{a DSS file reference from opendss}

  \item{rebuild}{Set to true to force rebuilding the DSS
  catalog file (.dsc).}
}
\value{
a character vector of DSS paths in the file.
}
\description{
Returns a list of all DSS paths in a file, useful for
searching for data.
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

