% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getColumnsByName}
\alias{getColumnsByName}
\title{getColumnsByName Get a column from a PairedDataContainer object.}
\usage{
getColumnsByName(file, pdc, column)
}
\value{
vector of values from column.
}
\description{
Gets a column from a paired data container by name.
}
\details{
Name of column must be exact or NA is returned.
}
\note{
NOTE
}
\author{
Evan Heisman
}

