% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{getDT}
\alias{getDT}
\title{getDT Get a TSC from a file and pathname as a data.table}
\usage{
getDT(file, path)
}
\value{
stuff
}
\description{
short desc
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Cameron Bracken
}
