\name{getFullDT}
\alias{getFullDT}
\title{getFullDT Get a full TSC as data.table, ignoring date parameters.}
\usage{
getFullDT(file, paths, discard_empty = TRUE)
}
\value{
merged data.tame of all times and values in the time series
matching paths.
}
\description{
Gets paths, converts to data.table, and merges to one time
series.
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Cameron Bracken
}

