% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{getLooseTSC}
\alias{getLooseTSC}
\title{getFullTSC Get a full TSC, ignoring date parameters.}
\usage{
getLooseTSC(file, paths, ...)
}
\value{
merged xts of all times and values in time series matching paths.
}
\description{
Gets paths, converts to XTS, and merges to one time series.
}
\details{
Warning - does not check that all paths are the same except for D part
}
\note{
NOTE
}
\seealso{
Other getTSC: 
\code{\link{getFullTSC}()},
\code{\link{getTSC}()}
}
\author{
Evan Heisman
}
\concept{getTSC}
