% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata get metadata from a tsc java object}
\usage{
getMetadata(tsc, colnamesSource = "parameter")
}
\value{
data.frame containing metadata
}
\description{
get metadata from a tsc java object
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

