% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedDataContainers.R
\name{getPairedDataAsDataFrame}
\alias{getPairedDataAsDataFrame}
\title{columnsToDataFrame}
\usage{
getPairedDataAsDataFrame(file, pdc, colnameFunction = joinParameterToLabel)
}
\arguments{
\item{file}{string of filename or HecDss object from `opendss`}

\item{pdc}{string of pathname or PairedDataContainer object}

\item{colnameFunction}{function that takes pdc object and metadata list to generate column names for the 'yOrdinates' from the PDC, default takes format of 'yParameter.yLabel'.}
}
\description{
columnsToDataFrame
}
