% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{getTimeSeriesAsDataFrame}
\alias{getTimeSeriesAsDataFrame}
\title{getTimeseriesAsDataFrame Get a TSC, ignoring date parameters.}
\usage{
getTimeSeriesAsDataFrame(file, path, leaveAsTibble = FALSE, ...)
}
\value{
data.frame of all times and values in time series matching paths.
}
\description{
Gets paths, converts to `data.frame`, er a `tibble` into a`base::data.frame`
}
\seealso{
Other getTSC: 
\code{\link{getFullTSC}()},
\code{\link{getTSC}()}
}
\author{
Evan Heisman
}
\concept{getTSC}
