% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTSC}
\alias{getTSC}
\title{getTSC Get a TSC from a file and pathname as a XTS.}
\usage{
getTSC(file, path, fullTSC = FALSE, ...)
}
\value{
xts from time series located at path in file.
}
\description{
Skips intermediate step of getting TimeSeriesContainer object.
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}
\seealso{
Other getTSC: \code{\link{getFullTSC}};
  \code{\link{getLooseTSC}}
}

