\name{initialize.dssrip}
\alias{initialize.dssrip}
\title{initialize.dssrip}
\usage{
  initialize.dssrip(as.package = F, dss_location = NULL,
    platform = NULL, jmemory = NULL)
}
\value{
  Nothing useful returned.
}
\description{
  Short description
}
\details{
  Long Description
}
\note{
  NOTE
}
\author{
  Evan Heisman
}

