\name{nash.sutcliffe}
\alias{nash.sutcliffe}
\title{Nash-Sutcliffe model coefficient
checks for better performance than assuming climatology or against another model.}
\usage{
nash.sutcliffe(x.obs, x.model, x.alt = mean(x.obs))
}
\description{
Nash-Sutcliffe model coefficient checks for better
performance than assuming climatology or against another
model.
}

