\name{pathByPartsRegex}
\alias{pathByPartsRegex}
\title{pathByPartsRegex Search path by parts using regex.}
\usage{
pathByPartsRegex(paths, pattern, pattern.parts = NULL)
}
\value{
stuff
}
\description{
Searches path by parts using regular expressions, e.g.
"A=.*CREEK.* C=FLOW"
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

