\name{pathByPartsWildcard}
\alias{pathByPartsWildcard}
\title{pathByPartsWildcard Search paths by parts, using wildcards.}
\usage{
pathByPartsWildcard(paths, pattern)
}
\value{
stuff
}
\description{
Searches path by individual parts, e.g. "A=*CREEK* C=FLOW"
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

