\name{rmse}
\alias{rmse}
\title{RMSE function, checked
one parameter assumes x1 is residuals,
two parameters assumes x1 and x2 are modeled versus fitted}
\usage{
rmse(x.obs, x.model = NULL)
}
\description{
RMSE function, checked one parameter assumes x1 is
residuals, two parameters assumes x1 and x2 are modeled
versus fitted
}

