\name{separatePathParts}
\alias{separatePathParts}
\title{separatePathParts Separates path parts into dataframe.}
\usage{
separatePathParts(paths)
}
\value{
data frame consisting of split path parts and full paths.
}
\description{
useful function for writing filters
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

