% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{squeeze}
\alias{squeeze}
\title{Squeeze a DSS file}
\usage{
squeeze(file)
}
\arguments{
\item{file}{dss file object from `opendss`}
}
\description{
Calls squeeze method on file to remove deleted or overwritten data
}
