% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{tsc.to.dt}
\alias{tsc.to.dt}
\title{tsc.to.dt Converts Java TimeSeriesContainer objects to data.table objects.}
\usage{
tsc.to.dt(tsc, ...)
}
\value{
data.table object from times and values in TSC.
}
\description{
convert time series container to data.table
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Cameron Bracken
}
