% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{tsc.to.xts}
\alias{tsc.to.xts}
\title{tsc.to.xts Converts Java TimeSeriesContainer objects into XTS time series objects.}
\usage{
tsc.to.xts(tsc, colnamesSource = "parameter", offsetForType = FALSE)
}
\arguments{
\item{colnamesSource}{which TimeSeriesContainer field should be used for column names, typically `parameter`}

\item{offsetForType}{defaults to `FALSE` for backwards compatibility, offsets data if period average}
}
\value{
xts object from times and values in TSC.
}
\description{
convert time series container to XTS
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}
