\docType{data}
\name{TSC_TYPES}
\alias{TSC_TYPES}
\title{Constants used in DSS
These are exported for building functions manipulating the raw rJava objects.}
\format{\preformatted{ chr [1:4] "INST-VAL" "INST-CUM" "PER-AVER" "PER-CUM"
}}
\usage{
TSC_TYPES
}
\description{
Constants used in DSS These are exported for building
functions manipulating the raw rJava objects.
}
\note{
Types for time series objects
}
\keyword{datasets}

