\name{wymonth}
\alias{wymonth}
\title{Month of water year}
\usage{
wymonth(t)
}
\arguments{
  \item{t}{timestamp object (e.g. Date, POSIXt)}
}
\value{
integer of month in water year (OCT = 1, SEP = 12) Works
similar to 'year' function on POSIXt classes
}
\description{
Get's the month in the water year (Oct=1, Sept=12)
}

