% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{xts.to.tsc}
\alias{xts.to.tsc}
\title{xts.to.tsc Converts xts objects to Java TimeSeriesContainer objects.}
\usage{
xts.to.tsc(
  tsObject,
  ...,
  ePart = NULL,
  interval = NULL,
  fillData = FALSE,
  protoTSC = NULL
)
}
\value{
java TimeSeriesContainer.
}
\description{
Converts xts objects to TimeSeriesContainers for writing to DSS files.
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}
