% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getLooseTSC}
\alias{getLooseTSC}
\title{getFullTSC Get a full TSC, ignoring date parameters.}
\usage{
getLooseTSC(file, paths, ...)
}
\value{
merged xts of all times and values in time series matching paths.
}
\description{
Gets paths, converts to XTS, and merges to one time series.
}
\details{
Warning - does not check that all paths are the same except for D part
}
\note{
NOTE
}
\author{
Evan Heisman
}
\seealso{
Other getTSC: \code{\link{getFullTSC}};
  \code{\link{getTSC}}
}

