% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesContainers.R
\name{getTSC}
\alias{getTSC}
\title{getTSC Get a TSC from a file and pathname as a XTS.}
\usage{
getTSC(file, path, fullTSC = FALSE, ...)
}
\value{
xts from time series located at path in file.
}
\description{
Skips intermediate step of getting TimeSeriesContainer object.
}
\details{
Long Description
}
\note{
NOTE
}
\seealso{
Other getTSC: 
\code{\link{getFullTSC}()}
}
\author{
Evan Heisman
}
\concept{getTSC}
