% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hydrologic_date_functions}
\alias{hydrologic_date_functions}
\alias{months14}
\alias{wateryear}
\alias{wyday}
\alias{wymonth}
\alias{wymonth.abb}
\title{Hydrologic date helper functions}
\usage{
wateryear(t)
}
\arguments{
\item{t}{timestamp object (e.g. \code{Date}, \code{POSIXt} classes)}

\item{splitMonths}{numerals for months that are to be split in two for 14 (or n) period models, defaults to April (7) and August (11).}

\item{splitDay}{day on which months should be split, defaults to 16.}

\item{monthAbbrevs}{the ordered abbreivations to use, defaults to wymonth.abb.  Use month.abb if you want calendar year order, month.name if you want full month names}

\item{sep}{separator between month and periodNames}

\item{case}{function to convert cases, \code{toupper} for upper case months, \code{tolower} for lower case, null (default) for no change}

\item{periodNames}{- naming convention for periods, defaults to 1 and 2, for example, could be I and II, or A and B.}
}
\value{
\code{wateryear} water year for a given timestamp (October of previous calendar year through September)

\code{wymonth} month of water year (Oct = 1, Sept = 12)

\code{wyday} day of water year (01Oct=1, 30Sep=365+leapyear(y))

\code{wymonth.abb} abbreviated months of year sorted by water year order

\code{months14} returns a function that when diven a timestamp, returns the 'abbreviated' 14 period months, can be more split periods than just typical 14-periods.
}
\description{
Provides 'lubridate' type functions for getting numeric values from `Date` and `POSIXt` objects based on the water year.
}
\note{
\code{wy} is no longer provided due to conflicts with the common use in hydrologic programs
}
\author{
Evan Heisman
}

