% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tsc.to.dt}
\alias{tsc.to.dt}
\title{tsc.to.dt Converts Java TimeSeriesContainer objects to data.table objects.}
\usage{
tsc.to.dt(tsc, ...)
}
\value{
data.table object from times andvalues in TSC.
}
\description{
convert time series container to data.table
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Cameron Bracken
}

