\name{xts.to.tsc}
\alias{xts.to.tsc}
\title{xts.to.tsc Converts xts objects to Java TimeSeriesContainer objects.}
\usage{
xts.to.tsc(tsObject, ..., protoTSC = NULL)
}
\value{
java TimeSeriesContainer.
}
\description{
Converts xts objects to TimeSeriesContainers for writing to
DSS files.
}
\details{
Long Description
}
\note{
NOTE
}
\author{
Evan Heisman
}

